in_tail: don't repeatedly add symlinks to scan queue
This commit stops paths that are actually symlinks from being
repeatedly added to the scan queue when the fluent-bit tail
input plugin is used on Windows.

The fix works by recognizing that there are actually two
different types of comparisons that we may sometimes wish
to do on filenames / paths. The first occurs when making a
determination of whether a filename / path is already on the
scan queueu or needs to be added. This should be a simple
comparison and whether the filename / path is for an actual
file or symlink is irrelevant. The second occurs when making
determinations about (for instance) whether a file on the
scan queue has been rotated-- in which case, if a filename /
path happens to be a symlink, we need to compare the
underlying target file instead.

This commit changes the logic of the flb_tail_file_name_cmp
function to have no regard for whether a filename / path is
a symlink or not to cover the first of the two cases
described above. It also introduces a new function--
flb_tail_target_file_name_cmp-- which can be used to cover
the second case described above. Where applicable, calls to
flb_tail_file_name_cmp were replaced with calls to
flb_tail_target_file_name_cmp.

Signed-off-by: Kent Rancourt <kent.rancourt@microsoft.com>
Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>